Moved tabPanel and temp drawer to navigation folder
